[device-bootloaders] Fixes to Asset Tracking Service to use Enum type for Command ID

+++ device-bootloaders/ServiceProcessorBL2/src/mailbox.c
@@ -181,10 +181,10 @@ static void mbox_task(void *pvParameters)
             } break;
             case GET_MODULE_MANUFACTURE_NAME...GET_MODULE_MEMORY_TYPE: {
                 // Process asset tracking service request cmd
-                asset_tracking_process_request(mbox, (uint32_t)*message_id); 
+                asset_tracking_process_request(mbox, (uint8_t)*message_id); 
 	    } break;
             default:
-                printf("Invalid message id: %" PRIu64 "\r\n", *message_id);
+                printf("Invalid message id: %" PRIu8 "\r\n", (uint8_t)*message_id);
                 printf("message length: %" PRIi64 ", buffer:\r\n", length);
                 for (int64_t i = 0; i < length; ++i) {
                     if (i % 8 == 0 && i != 0)
